requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-spectacular==0.16.0
drf-yasg==1.20.0
future==0.18.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
packaging==20.9
phonenumberslite==8.12.24
Pillow==8.2.0
pkg-resources==0.0.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
